requirements.txt
py==1.10.0
torch==1.7.0
torchvision==0.8.0
numpy==1.19.2
opencv-python==4.4.0.46
opencv-python-headless==4.5.3.56
pickleshare==0.7.5
Pillow==8.0.0
jenkspy==0.2.0